refactor: move language recipes into datasets
