+++ src/lerobot/datasets/language_render.py
@@ -21,11 +21,11 @@ import re
 from collections.abc import Sequence
 from typing import Any
 
-from lerobot.language.recipe import DEFAULT_BINDINGS, TrainingRecipe, render_message_turns
 from lerobot.utils.constants import LANGUAGE_PERSISTENT, MESSAGES_RENDERED
 from lerobot.utils.utils import unwrap_scalar
 
 from .language import column_for_style
+from .recipe import DEFAULT_BINDINGS, TrainingRecipe, render_message_turns
 
 LanguageRow = dict[str, Any]
 RenderedMessages = dict[str, list[Any]]
