+++ tests/processor/test_render_messages_processor.py
@@ -7,7 +7,7 @@ pytest.importorskip("datasets", reason="datasets is required (install lerobot[da
 import numpy as np  # noqa: E402
 import torch  # noqa: E402
 
-from lerobot.language.recipe import MessageTurn, TrainingRecipe  # noqa: E402
+from lerobot.datasets.recipe import MessageTurn, TrainingRecipe  # noqa: E402
 from lerobot.lerobot_types import TransitionKey  # noqa: E402
 from lerobot.processor.converters import create_transition  # noqa: E402
 from lerobot.processor.render_messages_processor import (  # noqa: E402
